% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customAnnot.R
\name{loadAnnot}
\alias{loadAnnot}
\title{Load Compound Annotation Database}
\usage{
loadAnnot()
}
\description{
The compound annotation tables from different databases/sources are stored
in one SQLite database. This function can be used to load the SQLite 
annotation database
}
\examples{
conn <- loadAnnot()
}
