% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugage_annot.R
\name{processDrugage}
\alias{processDrugage}
\title{Process Source DrugAge Dataset}
\usage{
processDrugage(
  drugagefile = "drugage_id_mapping.tsv",
  redownloaddrugage = FALSE
)
}
\arguments{
\item{drugagefile}{character(1), path to the destination DrugAge annotation 
file with id mappings}

\item{redownloaddrugage}{TRUE or FALSE indicating whether to redownload the 
DrugAge dataset if it has been downloaded before}
}
\value{
write the 'drugage_id_mapping.tsv' table
}
\description{
This function processes the source DrugAge datasets by adding the ChEMBL, 
PubChem and DrugBank id mapping information to the source DrugAge table 
which only has comopund names without id mapping information. Source file of 
DrugAge is linked here: \url{http://genomics.senescence.info/drugs/dataset.zip}
}
\details{
This function only annotates compound names that have ChEMBL 
ids from the ChEMBL database (version 24). The missing IDs were added 
manually. A semi-manual approach was to use this web service: 
\url{https://cts.fiehnlab.ucdavis.edu/batch}. After the semi-manual process,
the left ones were manually mapped to ChEMBL, PubChem and DrugBank ids. 
The mixed items were commented.
}
\examples{
library(ChemmineR)
\dontrun{
processDrugage(drugagefile="drugage_id_mapping.tsv", redownloaddrugage=FALSE)
# Now the missing IDs need to be added manually. A semi-manual approach is to 
# use this web service: https://cts.fiehnlab.ucdavis.edu/batch
}
}
